feat: add `HasContent()` with more options for `IFileInfo`
