+++ Tests/aweXpect.Testably.Tests/FileSystem.HasFile.WithContent.Tests.cs
@@ -4,12 +4,56 @@ using Testably.Abstractions.Testing;
 
 namespace aweXpect.Testably.Tests;
 
-public sealed partial class HasFile
+public sealed partial class FileSystem
 {
-	public sealed partial class WithContent
+	public sealed partial class HasFile
 	{
-		public class EqualTo
+		public sealed partial class WithContent
 		{
+			public sealed class Tests
+			{
+				[Fact]
+				public async Task WhenContentIsDifferent_ShouldFail()
+				{
+					string path = "foo.txt";
+					IFileSystem sut = new MockFileSystem();
+					// ReSharper disable once MethodHasAsyncOverload
+					sut.File.WriteAllText(path, "baz");
+
+					async Task Act()
+						=> await That(sut).HasFile(path).WithContent("bar");
+
+					await That(Act).ThrowsException()
+						.WithMessage($"""
+						              Expected that sut
+						              has file '{path}' with content equal to "bar",
+						              but it was "baz" which differs at index 2:
+						                   ↓ (actual)
+						                "baz"
+						                "bar"
+						                   ↑ (expected)
+
+						              File content:
+						              baz
+						              """);
+				}
+
+				[Fact]
+				public async Task WhenContentMatches_ShouldSucceed()
+				{
+					string path = "foo.txt";
+					string content = "bar";
+					IFileSystem sut = new MockFileSystem();
+					// ReSharper disable once MethodHasAsyncOverload
+					sut.File.WriteAllText(path, content);
+
+					async Task Act()
+						=> await That(sut).HasFile(path).WithContent(content);
+
+					await That(Act).DoesNotThrow();
+				}
+			}
+
 			public sealed class BinaryTests
 			{
 				[Fact]
@@ -23,7 +67,7 @@ public sealed partial class HasFile
 					sut.File.WriteAllBytes(path, content);
 
 					async Task Act()
-						=> await That(sut).HasFile(path).WithContent().EqualTo(expected);
+						=> await That(sut).HasFile(path).WithContent(expected);
 
 					await That(Act).ThrowsException()
 						.WithMessage($"""
@@ -43,51 +87,7 @@ public sealed partial class HasFile
 					sut.File.WriteAllBytes(path, content);
 
 					async Task Act()
-						=> await That(sut).HasFile(path).WithContent().EqualTo(content);
-
-					await That(Act).DoesNotThrow();
-				}
-			}
-
-			public sealed class StringTests
-			{
-				[Fact]
-				public async Task WhenContentIsDifferent_ShouldFail()
-				{
-					string path = "foo.txt";
-					IFileSystem sut = new MockFileSystem();
-					// ReSharper disable once MethodHasAsyncOverload
-					sut.File.WriteAllText(path, "baz");
-
-					async Task Act()
-						=> await That(sut).HasFile(path).WithContent().EqualTo("bar");
-
-					await That(Act).ThrowsException()
-						.WithMessage($"""
-						              Expected that sut
-						              has file '{path}' with content equal to "bar",
-						              but it was "baz" which differs at index 2:
-						                   ↓ (actual)
-						                "baz"
-						                "bar"
-						                   ↑ (expected)
-						              
-						              File content:
-						              baz
-						              """);
-				}
-
-				[Fact]
-				public async Task WhenContentMatches_ShouldSucceed()
-				{
-					string path = "foo.txt";
-					string content = "bar";
-					IFileSystem sut = new MockFileSystem();
-					// ReSharper disable once MethodHasAsyncOverload
-					sut.File.WriteAllText(path, content);
-
-					async Task Act()
-						=> await That(sut).HasFile(path).WithContent().EqualTo(content);
+						=> await That(sut).HasFile(path).WithContent(content);
 
 					await That(Act).DoesNotThrow();
 				}
@@ -104,7 +104,7 @@ public sealed partial class HasFile
 					sut.File.WriteAllText(path, "baz");
 
 					async Task Act()
-						=> await That(sut).HasFile(path).WithContent().EqualTo("b?").AsWildcard();
+						=> await That(sut).HasFile(path).WithContent("b?").AsWildcard();
 
 					await That(Act).ThrowsException()
 						.WithMessage($"""
@@ -115,7 +115,7 @@ public sealed partial class HasFile
 						                "baz"
 						                "b?"
 						                ↑ (wildcard pattern)
-						              
+
 						              File content:
 						              baz
 						              """);
@@ -130,7 +130,7 @@ public sealed partial class HasFile
 					sut.File.WriteAllText(path, "bar");
 
 					async Task Act()
-						=> await That(sut).HasFile(path).WithContent().EqualTo("ba?").AsWildcard();
+						=> await That(sut).HasFile(path).WithContent("ba?").AsWildcard();
 
 					await That(Act).DoesNotThrow();
 				}
